move functions to proper files, fmt, lint, test, proto...
